+++ rpc/src/v1/types/get_block_response.rs
@@ -47,7 +47,7 @@ pub struct VerboseBlock {
 	/// Block nbits
 	pub bits: u32,
 	/// Block difficulty
-	pub difficulty: u32,
+	pub difficulty: f64,
 	/// Expected number of hashes required to produce the chain up to this block (in hex)
 	pub chainwork: U256,
 	/// Hash of previous block
